add spawn mongocryptd spec to client spec
