+++ spec/mongo/client_spec.rb
@@ -842,6 +842,64 @@ describe Mongo::Client do
   end
 
   describe '#spawn_mongocryptd' do
-   
+    require_libmongocrypt
+    require_enterprise
+
+    let(:client) do
+      new_local_client_nmio(
+        [SpecConfig.instance.addresses.first],
+        {
+          auto_encryption_options: {
+            key_vault_client: new_local_client_nmio('mongodb://127.0.0.1:27018'),
+            key_vault_namespace: 'database.collection',
+            kms_providers: {
+              local: { key: Base64.encode64('ruby' * 24) },
+            },
+            extra_options: extra_options
+          }
+        }
+      )
+    end
+
+    context 'without auto encryption options' do
+      let(:client) { new_local_client_nmio([SpecConfig.instance.addresses.first]) }
+
+      it 'raises an exception' do
+        expect { client.spawn_mongocryptd }.to raise_error(ArgumentError, 'Cannot spawn mongocryptd process without setting auto encryption options on the client.')
+      end
+    end
+
+    context 'with auto encryption options' do
+      let(:extra_options) { {} }
+      let(:pid) { client.spawn_mongocryptd }
+
+      after do
+        Process.kill('TERM', pid)
+        Process.wait(pid)
+      end
+
+      it 'spawns mongocryptd' do
+        # Verify that the process at pid is still running -
+        # every active process will have a process group, so
+        # if the process group id is a number, the process is
+        # still running
+        expect(Process.getpgid(pid)).to be_a_kind_of(Numeric)
+      end
+    end
+
+    context 'with empty arguments and shell command as path' do
+      let(:extra_options) do
+        {
+          mongocryptd_spawn_path: 'echo hello world',
+          mongocryptd_spawn_args: []
+        }
+      end
+
+      it 'raises an exception' do
+        expect do
+          client.spawn_mongocryptd
+        end.to raise_error(Errno::ENOENT, /No such file or directory - echo hello world/)
+      end
+    end
   end
 end
